Complete Integration
Import existing differences into kernel to fully merge the external
build and the internal version to create consistent results regardless
of whether a kernel was installed post build or as a part of the
compile process. - This should resolve any remaining issues that a
separate kernel install would have successfully eliminated in the past.